feat: add location caching

+++ app/src/main/java/com/example/up/data/local/mappers/PositionMapper.kt
@@ -5,7 +5,7 @@ import com.example.up.domain.model.Position
 
 fun Location.toDomain(): Position{
     return Position(
-        lat = this.latitude.toFloat(),
-        lon = this.longitude.toFloat()
+        lat = this.latitude,
+        lon = this.longitude
     )
 }
